SingleCollege Implemented
